+++ packages/core/src/files/Helpers.ts
@@ -1,5 +1,5 @@
 /*
-Copyright 2021-2022 Javier Brea
+Copyright 2021-2023 Javier Brea
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
 
@@ -8,27 +8,30 @@ http://www.apache.org/licenses/LICENSE-2.0
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
 
-const path = require("path");
+import path from "path";
 
-const fsExtra = require("fs-extra");
-const yaml = require("yaml");
-const { flatten, uniq } = require("lodash");
+import type { RegisterOptions } from "@babel/register";
+import { readFile } from "fs-extra";
+import { flatten, uniq } from "lodash";
+import yaml from "yaml";
+
+import type { FileLoaded, ErrorLoadingFile } from "./FilesLoader.types";
 
 const YAML_EXTENSIONS = [".yaml", ".yml"];
 const DEFAULT_EXTENSIONS = [".json", ".js", ".cjs"].concat(YAML_EXTENSIONS);
 const BABEL_DEFAULT_EXTENSIONS = [".es6", ".es", ".esm", ".cjs", ".jsx", ".js", ".mjs", ".ts"];
 
-function globuleExtensionPattern(srcGlob, extension) {
+function globuleExtensionPattern(srcGlob: string, extension: string): string {
   return `${srcGlob}${extension}`;
 }
 
-function extensionsGlobulePatterns(srcGlob, extensions) {
+function extensionsGlobulePatterns(srcGlob: string, extensions: string[]): string[] {
   return extensions.map((extension) => {
     return globuleExtensionPattern(srcGlob, extension);
   });
 }
 
-function getGlobulePatterns(src, extensions) {
+function getGlobulePatterns(src: string, extensions: string[]): string[] {
   const srcs = Array.isArray(src) ? src : [src];
   return uniq(
     flatten(
@@ -39,8 +42,11 @@ function getGlobulePatterns(src, extensions) {
   );
 }
 
-function getFilesExtensions(babelRegister, babelRegisterOptions) {
-  if (!!babelRegister) {
+function getFilesExtensions(
+  babelRegister: boolean,
+  babelRegisterOptions: RegisterOptions
+): string[] {
+  if (babelRegister) {
     if (babelRegisterOptions.extensions) {
       return DEFAULT_EXTENSIONS.concat(babelRegisterOptions.extensions);
     }
@@ -49,17 +55,26 @@ function getFilesExtensions(babelRegister, babelRegisterOptions) {
   return DEFAULT_EXTENSIONS;
 }
 
-function getFilesGlobule(src, babelRegister, babelRegisterOptions) {
+export function getFilesGlobule(
+  src: string,
+  babelRegister: boolean,
+  babelRegisterOptions: RegisterOptions
+) {
   return getGlobulePatterns(src, uniq(getFilesExtensions(babelRegister, babelRegisterOptions)));
 }
 
-function babelRegisterOnlyFilter(collectionsFolder) {
-  return (filePath) => {
-    return filePath.indexOf(collectionsFolder) === 0;
+export function babelRegisterOnlyFilter(
+  collectionsFolder: string
+): (filePath?: string) => boolean {
+  return (filePath?: string): boolean => {
+    return filePath ? filePath.indexOf(collectionsFolder) === 0 : false;
   };
 }
 
-function babelRegisterDefaultOptions(collectionsFolder, babelRegisterOptions) {
+export function babelRegisterDefaultOptions(
+  collectionsFolder: string,
+  babelRegisterOptions: RegisterOptions
+): RegisterOptions {
   return {
     only: [babelRegisterOnlyFilter(collectionsFolder)],
     cache: false,
@@ -68,27 +83,27 @@ function babelRegisterDefaultOptions(collectionsFolder, babelRegisterOptions) {
   };
 }
 
-function validateFileContent(fileContent) {
+export function validateFileContent(fileContent: unknown): string | null {
   if (!Array.isArray(fileContent)) {
     return "File does not export an array";
   }
   return null;
 }
 
-function isYamlFile(filePath) {
+export function isYamlFile(filePath: string): boolean {
   return YAML_EXTENSIONS.includes(path.extname(filePath));
 }
 
-function readYamlFile(filePath) {
-  return fsExtra.readFile(filePath, "utf8").then((fileContent) => {
+export function readYamlFile(filePath: string): Promise<unknown> {
+  return readFile(filePath, "utf8").then((fileContent) => {
     return yaml.parse(fileContent);
   });
 }
 
-module.exports = {
-  babelRegisterDefaultOptions,
-  getFilesGlobule,
-  validateFileContent,
-  isYamlFile,
-  readYamlFile,
-};
+export function isFileLoaded(file: FileLoaded | ErrorLoadingFile): file is FileLoaded {
+  return (file as FileLoaded).content !== undefined;
+}
+
+export function isErrorLoadingFile(file: FileLoaded | ErrorLoadingFile): file is ErrorLoadingFile {
+  return (file as ErrorLoadingFile).error !== undefined;
+}