+++ packages/core/test/components/files/Files.spec.js
@@ -1,5 +1,5 @@
 /*
-Copyright 2019 Javier Brea
+Copyright 2023 Javier Brea
 Copyright 2019 XbyOrange
 
 Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License. You may obtain a copy of the License at
@@ -8,6 +8,7 @@ http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 */
+jest.mock("fs-extra");
 
 const path = require("path");
 const sinon = require("sinon");
@@ -21,7 +22,7 @@ const LibsMocks = require("../common/Libs.mocks");
 const CoreMocks = require("../Core.mocks");
 const ConfigMock = require("../common/Config.mocks");
 
-const FilesLoaders = require("../../../src/files/FilesLoaders");
+const { Files } = require("../../../src/files/Files");
 
 const wait = () => {
   return new Promise((resolve) => {
@@ -31,7 +32,7 @@ const wait = () => {
   });
 };
 
-describe("FilesLoaders", () => {
+describe("Files", () => {
   const fooRequireCache = {
     "foo-path": {
       id: "foo-path",
@@ -67,7 +68,7 @@ describe("FilesLoaders", () => {
   let coreMocks;
   let coreInstance;
   let requireCache;
-  let filesLoader;
+  let files;
   let libsMocks;
   let pluginMethods;
   let pathOption;
@@ -105,13 +106,14 @@ describe("FilesLoaders", () => {
       logger,
     };
 
-    filesLoader = new FilesLoaders(pluginMethods, {
+    files = new Files(pluginMethods, {
       requireCache,
     });
     sandbox.stub(path, "isAbsolute").returns(true);
     sandbox.stub(console, "log");
     sandbox.stub(yaml, "parse");
-    libsMocks.stubs.fsExtra.existsSync.returns(true);
+    jest.spyOn(fsExtra, "existsSync").mockImplementation(() => true);
+    jest.spyOn(fsExtra, "readFile").mockResolvedValue({});
     libsMocks.stubs.globule.find.returns([]);
     pathOption = { value: "foo-path" };
     watchOption = { value: true };
@@ -119,15 +121,15 @@ describe("FilesLoaders", () => {
     babelRegisterOptionsOption = { value: {} };
     enabledOption = { value: true };
 
-    filesLoader._enabledOption = enabledOption;
-    filesLoader._pathOption = pathOption;
-    filesLoader._watchOption = watchOption;
-    filesLoader._babelRegisterOption = babelRegisterOption;
-    filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
+    files._enabledOption = enabledOption;
+    files._pathOption = pathOption;
+    files._watchOption = watchOption;
+    files._babelRegisterOption = babelRegisterOption;
+    files._babelRegisterOptionsOption = babelRegisterOptionsOption;
   });
 
   afterEach(async () => {
-    await filesLoader.stop();
+    await files.stop();
     libsMocks.restore();
     sandbox.restore();
     coreMocks.restore();
@@ -135,36 +137,36 @@ describe("FilesLoaders", () => {
 
   describe("id", () => {
     it("should return files", async () => {
-      expect(FilesLoaders.id).toEqual("files");
+      expect(Files.id).toEqual("files");
     });
   });
 
   describe("path getter", () => {
     it("should return the resolved value of the path option", async () => {
       path.isAbsolute.returns(false);
-      expect(filesLoader.path).toEqual(path.resolve(process.cwd(), "foo-path"));
+      expect(files.path).toEqual(path.resolve(process.cwd(), "foo-path"));
     });
   });
 
   describe("loaders getter", () => {
     it("should return loaders", async () => {
-      await filesLoader.init();
-      expect(filesLoader.loaders.collections).toBeDefined();
-      expect(filesLoader.loaders.routes).toBeDefined();
+      await files.init();
+      expect(files.loaders.collections).toBeDefined();
+      expect(files.loaders.routes).toBeDefined();
     });
   });
 
   describe("when initialized", () => {
     it("should not require files from mocks folders if it is disabled", async () => {
       enabledOption.value = false;
-      await filesLoader.init();
+      await files.init();
       expect(libsMocks.stubs.fsExtra.ensureDirSync.callCount).toEqual(0);
     });
 
     it("should require all files from mocks folders calculating it from cwd if path is not absolute", async () => {
       path.isAbsolute.returns(false);
-      libsMocks.stubs.fsExtra.existsSync.returns(false);
-      await filesLoader.init();
+      fsExtra.existsSync.mockReturnValue(false);
+      await files.init();
       expect(libsMocks.stubs.fsExtra.ensureDirSync.getCall(0).args[0]).toEqual(
         path.resolve(process.cwd(), "foo-path")
       );
@@ -172,21 +174,21 @@ describe("FilesLoaders", () => {
 
     it("should not throw and add an alert if there is an error loading route files", async () => {
       libsMocks.stubs.globule.find.returns(["foo"]);
-      await filesLoader.init();
+      await files.init();
       expect(alerts.flat.length).toEqual(1);
     });
 
     it("should add an alert when a routes file content does not pass validation", async () => {
       libsMocks.stubs.globule.find.returns(["foo"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => ({}),
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat[0].id).toEqual("files:loader:routes:file:foo");
       expect(alerts.flat[0].message).toEqual(
         expect.stringContaining("Error loading routes from file foo")
@@ -195,15 +197,15 @@ describe("FilesLoaders", () => {
 
     it("should add an alert when collections file content does not pass validation", async () => {
       libsMocks.stubs.globule.find.returns(["collections.json"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => ({}),
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat[0].id).toEqual("files:loader:collections:error");
       expect(alerts.flat[0].message).toEqual(
         expect.stringContaining("Error loading collections from file")
@@ -212,63 +214,85 @@ describe("FilesLoaders", () => {
 
     it("should not add an alert when a routes file content pass validation", async () => {
       libsMocks.stubs.globule.find.returns(["foo"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat.length).toEqual(0);
     });
 
     it("should not add an alert when a routes file content is yaml and pass validation", async () => {
       yaml.parse.returns([]);
       libsMocks.stubs.globule.find.returns(["foo.yml"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat.length).toEqual(0);
     });
 
+    it("should add an alert if there is an error processing files in loaders", async () => {
+      yaml.parse.returns([]);
+      libsMocks.stubs.globule.find.returns(["foo.yml"]);
+      files = new Files(pluginMethods, {
+        requireCache,
+        require: () => [],
+      });
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
+        id: "foo",
+        src: "foo",
+        onLoad: () => {
+          throw new Error("Foo loader error");
+        },
+      });
+      await files.init();
+      expect(alerts.flat[0].error.message).toEqual("Foo loader error");
+    });
+
     it("should not throw and add an alert if there is an error loading mocks file", async () => {
-      await filesLoader.init();
+      await files.init();
       expect(alerts.flat.length).toEqual(1);
     });
 
     it("should remove alerts when collections file loads successfully", async () => {
       libsMocks.stubs.globule.find.returns(["foo"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat.length).toEqual(0);
     });
 
     it("should add an alert when no collections file is found", async () => {
       libsMocks.stubs.globule.find.returns([]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat.length).toEqual(1);
       expect(alerts.flat[0].message).toEqual(
         expect.stringContaining("No collections file was found")
@@ -276,29 +300,29 @@ describe("FilesLoaders", () => {
     });
 
     it("should call to loadCollections method when mocks file is loaded", async () => {
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(pluginMethods.loadCollections.callCount).toEqual(1);
     });
 
     it("should add an alert when file name is mocks", async () => {
       libsMocks.stubs.globule.find.returns(["mocks.json"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat[0].message).toEqual(
         "Defining collections in 'mocks.json' file is deprecated. Please rename it to 'collections.json'"
       );
@@ -306,37 +330,37 @@ describe("FilesLoaders", () => {
 
     it("should remove alert when file name is collections", async () => {
       libsMocks.stubs.globule.find.returns(["mocks.json"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => [],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(alerts.flat[0].message).toEqual(
         "Defining collections in 'mocks.json' file is deprecated. Please rename it to 'collections.json'"
       );
       libsMocks.stubs.globule.find.returns(["collections.json"]);
-      await filesLoader.reload();
+      await files.reload();
       expect(alerts.flat.length).toEqual(0);
     });
 
     it("should not throw and add an alert if there is an error in loadRoutesfiles method", async () => {
-      sandbox.stub(filesLoader, "_readFile").throws(new Error());
-      await filesLoader.init();
+      sandbox.stub(files, "_readFile").throws(new Error());
+      await files.init();
       expect(alerts.flat.length).toEqual(1);
     });
 
     it("should return a rejected promise if there is an error initializing", async () => {
       Logger.prototype.info.throws(new Error("foo error"));
-      await expect(() => filesLoader.init()).rejects.toThrow("foo error");
+      await expect(() => files.init()).rejects.toThrow("foo error");
     });
 
     it("should do nothing if folder exists", async () => {
-      libsMocks.stubs.fsExtra.existsSync.returns(true);
-      await filesLoader.init();
+      fsExtra.existsSync.mockReturnValue(true);
+      await files.init();
       expect(libsMocks.stubs.fsExtra.ensureDirSync.callCount).toEqual(0);
     });
 
@@ -345,15 +369,15 @@ describe("FilesLoaders", () => {
         //do nothing
       });
       coreInstance.lowLevelConfig = { babelRegister: true, babelRegisterOptions: {} };
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: requireSpy,
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = { value: true };
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      await filesLoader.init();
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = { value: true };
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      await files.init();
       expect(requireSpy.getCall(0).args[0]).toEqual("@babel/register");
     });
 
@@ -361,35 +385,35 @@ describe("FilesLoaders", () => {
       const fooCachePath = "foo-path";
 
       expect(requireCache[fooCachePath]).toBeDefined();
-      await filesLoader.init();
+      await files.init();
       expect(requireCache[fooCachePath]).not.toBeDefined();
     });
 
     it("should require cache in order to found the mocks folder", async () => {
-      filesLoader = new FilesLoaders(pluginMethods);
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      sandbox.spy(filesLoader, "_cleanRequireCache");
-      await filesLoader.init();
+      files = new Files(pluginMethods);
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      sandbox.spy(files, "_cleanRequireCache");
+      await files.init();
       // it seems like require cache is empty in jest environment
-      expect(filesLoader._cleanRequireCache.callCount).toEqual(0);
+      expect(files._cleanRequireCache.callCount).toEqual(0);
     });
 
-    it("should clean require cache for mocks folder childs", async () => {
+    it("should clean require cache for mocks folder children", async () => {
       const fooCachePath = "foo-path/foo-children";
 
       expect(requireCache[fooCachePath]).toBeDefined();
-      await filesLoader.init();
+      await files.init();
       expect(requireCache[fooCachePath]).not.toBeDefined();
     });
 
-    it("should clean require cache for mocks folder childs recursively", async () => {
+    it("should clean require cache for mocks folder children recursively", async () => {
       const fooCachePath = "foo-path/foo-children-2";
 
       expect(requireCache[fooCachePath]).toBeDefined();
-      await filesLoader.init();
+      await files.init();
       expect(requireCache[fooCachePath]).not.toBeDefined();
     });
   });
@@ -397,24 +421,24 @@ describe("FilesLoaders", () => {
   describe("start method", () => {
     describe("when starting files watch", () => {
       it("should do nothing if files was not enabled", async () => {
-        filesLoader._enabledOption.value = false;
-        await filesLoader.init();
-        await filesLoader.start();
+        files._enabledOption.value = false;
+        await files.init();
+        await files.start();
         expect(libsMocks.stubs.watch.callCount).toEqual(0);
       });
 
       it("should do nothing if watch was not enabled", async () => {
-        filesLoader._watchOption.value = false;
-        await filesLoader.init();
-        await filesLoader.start();
+        files._watchOption.value = false;
+        await files.init();
+        await files.start();
         expect(libsMocks.stubs.watch.callCount).toEqual(0);
       });
 
       it("should call to close watcher if watch was enabled previously", async () => {
-        filesLoader._watchOption.value = true;
-        await filesLoader.init();
-        await filesLoader.start();
-        await filesLoader.start();
+        files._watchOption.value = true;
+        await files.init();
+        await files.start();
+        await files.start();
         expect(libsMocks.stubs.watchClose.callCount).toEqual(1);
       });
     });
@@ -422,31 +446,31 @@ describe("FilesLoaders", () => {
 
   describe("when a file is changed", () => {
     it("should load files again", async () => {
-      sandbox.stub(filesLoader, "_loadFiles");
-      filesLoader._watchOption.value = true;
-      await filesLoader.init();
-      await filesLoader.start();
+      sandbox.stub(files, "_loadFiles");
+      files._watchOption.value = true;
+      await files.init();
+      await files.start();
       libsMocks.stubs.watch.getCall(0).args[2]();
       await wait();
-      expect(filesLoader._loadFiles.callCount).toEqual(2);
+      expect(files._loadFiles.callCount).toEqual(2);
     });
   });
 
   describe("when core settings change", () => {
     it("should enable watch again if path setting is changed", async () => {
-      filesLoader._watchOption.value = true;
-      await filesLoader.init();
-      await filesLoader.start();
+      files._watchOption.value = true;
+      await files.init();
+      await files.start();
       configMock.stubs.option.onChange.getCall(0).args[0]("foo-path-2");
       await wait();
       expect(libsMocks.stubs.watch.callCount).toEqual(2);
     });
 
     it("should disable watch if watch is changed", async () => {
-      filesLoader._watchOption.value = true;
-      await filesLoader.init();
-      await filesLoader.start();
-      filesLoader._watchOption.value = false;
+      files._watchOption.value = true;
+      await files.init();
+      await files.start();
+      files._watchOption.value = false;
       configMock.stubs.option.onChange.getCall(1).args[0](false);
       await wait();
       expect(libsMocks.stubs.watchClose.callCount).toEqual(1);
@@ -457,29 +481,29 @@ describe("FilesLoaders", () => {
     it("should call to its load function when files are loaded", async () => {
       libsMocks.stubs.globule.find.returns(["foo/foo-path/**"]);
       const spy = sandbox.spy();
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => ["foo-content"],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
         onLoad: spy,
       });
-      await filesLoader.init();
-      expect(libsMocks.stubs.globule.find.getCall(0).args[0]).toEqual({
+      await files.init();
+      expect(libsMocks.stubs.globule.find.getCall(0).args[0]).toEqual([
+        "foo/foo-path/**/*.json",
+        "foo/foo-path/**/*.js",
+        "foo/foo-path/**/*.cjs",
+        "foo/foo-path/**/*.yaml",
+        "foo/foo-path/**/*.yml",
+      ]);
+      expect(libsMocks.stubs.globule.find.getCall(0).args[1]).toEqual({
         prefixBase: true,
-        src: [
-          "foo/foo-path/**/*.json",
-          "foo/foo-path/**/*.js",
-          "foo/foo-path/**/*.cjs",
-          "foo/foo-path/**/*.yaml",
-          "foo/foo-path/**/*.yml",
-        ],
         srcBase: "foo-path",
       });
       expect(spy.getCall(0).args[0]).toEqual([
@@ -490,20 +514,20 @@ describe("FilesLoaders", () => {
     it("should call to its load function passing the result when file exports a function", async () => {
       libsMocks.stubs.globule.find.returns(["foo/foo-file.js"]);
       const spy = sandbox.spy();
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => () => ["foo-content"],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
         onLoad: spy,
       });
-      await filesLoader.init();
+      await files.init();
       expect(spy.getCall(0).args[0]).toEqual([
         { path: "foo/foo-file.js", content: ["foo-content"] },
       ]);
@@ -519,20 +543,20 @@ describe("FilesLoaders", () => {
         });
       };
       const spy = sandbox.spy();
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => exportedFunction,
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
         onLoad: spy,
       });
-      await filesLoader.init();
+      await files.init();
       expect(spy.getCall(0).args[0]).toEqual([
         { path: "foo/foo-file.js", content: ["foo-promise-content"] },
       ]);
@@ -549,20 +573,20 @@ describe("FilesLoaders", () => {
         });
       };
       const spy = sandbox.spy();
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => exportedFunction,
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
         onLoad: spy,
       });
-      await filesLoader.init();
+      await files.init();
       expect(spy.getCall(0).args[0]).toEqual([]);
       expect(spy.getCall(0).args[1]).toEqual([{ path: "foo/foo-file.js", error: promiseError }]);
     });
@@ -570,27 +594,27 @@ describe("FilesLoaders", () => {
     it("should support async onLoad functions", async () => {
       libsMocks.stubs.globule.find.returns(["foo/foo-path/**"]);
       const spy = sandbox.spy();
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => ["foo-content"],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
-        onLoad: (files) => {
+        onLoad: (loadedFiles) => {
           return new Promise((resolve) => {
             setTimeout(() => {
-              spy(files);
+              spy(loadedFiles);
               resolve();
             }, 200);
           });
         },
       });
-      await filesLoader.init();
+      await files.init();
       expect(spy.getCall(0).args[0]).toEqual([
         { path: "foo/foo-path/**", content: ["foo-content"] },
       ]);
@@ -598,15 +622,15 @@ describe("FilesLoaders", () => {
 
     it("should catch error and add an alert if onLoad promise is rejected", async () => {
       libsMocks.stubs.globule.find.returns(["foo/foo-path/**"]);
-      filesLoader = new FilesLoaders(pluginMethods, {
+      files = new Files(pluginMethods, {
         requireCache,
         require: () => ["foo-content"],
       });
-      filesLoader._pathOption = pathOption;
-      filesLoader._watchOption = watchOption;
-      filesLoader._babelRegisterOption = babelRegisterOption;
-      filesLoader._babelRegisterOptionsOption = babelRegisterOptionsOption;
-      filesLoader.createLoader({
+      files._pathOption = pathOption;
+      files._watchOption = watchOption;
+      files._babelRegisterOption = babelRegisterOption;
+      files._babelRegisterOptionsOption = babelRegisterOptionsOption;
+      files.createLoader({
         id: "foo",
         src: "foo/foo-path/**/*",
         onLoad: () => {
@@ -617,8 +641,8 @@ describe("FilesLoaders", () => {
           });
         },
       });
-      await filesLoader.init();
-      expect(alerts.flat[0].message).toEqual("Error proccesing loaded files");
+      await files.init();
+      expect(alerts.flat[0].message).toEqual("Error processing loaded files");
       expect(alerts.flat[0].error.message).toEqual("foo-error");
     });
   });